CRUD de imagens e gerenciamento dos arquivos de imagens

+++ App/Models/JogoCRUD.php
@@ -23,6 +23,22 @@
             try {
                 $this->pdo->beginTransaction();
 
+                $sqlExistente = "
+                    SELECT COUNT(*) FROM jogo
+                    WHERE titulo = :titulo AND plataforma = :plataforma
+                ";
+                $stmtExistente = $this->pdo->prepare(query: $sqlExistente);
+                $stmtExistente->bindValue(param: ':titulo', value: $jogo->getTitulo(), type: PDO::PARAM_STR);
+                $stmtExistente->bindValue(param: ':plataforma', value: $jogo->getPlataforma(), type: PDO::PARAM_STR);
+                $stmtExistente->execute();
+
+                $existe = (bool) $stmtExistente->fetchColumn();
+
+                if ($existe) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Jogo já existe.");
+                }
+
                 $sqlJogo = "
                     INSERT INTO jogo (
                         titulo, plataforma, data_lancamento, desenvolvedora, link_compra, descricao
@@ -37,7 +53,12 @@
                 $stmtJogo->bindValue(param: ':desenvolvedora', value: $jogo->getDesenvolvedora(), type: PDO::PARAM_STR);
                 $stmtJogo->bindValue(param: ':link',           value: $jogo->getLink(),           type: PDO::PARAM_STR);
                 $stmtJogo->bindValue(param: ':descricao',      value: $jogo->getDescricao(),      type: PDO::PARAM_STR);
-                $stmtJogo->execute();
+                $sucesso = $stmtJogo->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao inserir jogo.");
+                }
 
                 // ultimo id inserido no BD
                 $idJogo = (int) $this->pdo->lastInsertId();
@@ -47,10 +68,15 @@
                 $stmtJG = $this->pdo->prepare(query: $sqlJG);
                 $stmtJG->bindValue(param: ':id_jogo',   value: $idJogo,              type: PDO::PARAM_INT);
                 $stmtJG->bindValue(param: ':id_genero', value: $jogo->GetGenero(),  type: PDO::PARAM_INT);
-                $stmtJG->execute();
+                $sucesso =$stmtJG->execute();
 
-                // Define Logo Padrão
-                $sqlLogo = "
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao inserir gênero.");
+                }
+
+                // Define poster Padrão
+                $sqlPoster = "
                     INSERT INTO jogo_imagem (
                         id_jogo, 
                         caminho,
@@ -65,12 +91,17 @@
                     )
                 ";
 
-                $stmtLogo = $this->pdo->prepare(query: $sqlLogo);
-                $stmtLogo->bindValue(param: ':id_jogo', value: $idJogo, type: PDO::PARAM_INT);
-                $stmtLogo->bindValue(param: ':caminho', value: 'img/logo.png', type: PDO::PARAM_STR);
-                $stmtLogo->bindValue(param: ':tipo', value: 'logo', type: PDO::PARAM_STR);
-                $stmtLogo->bindValue(param: ':ordem_exib', value: 0, type: PDO::PARAM_INT);
-                $stmtLogo->execute();
+                $stmtPoster = $this->pdo->prepare(query: $sqlPoster);
+                $stmtPoster->bindValue(param: ':id_jogo', value: $idJogo, type: PDO::PARAM_INT);
+                $stmtPoster->bindValue(param: ':caminho', value: '/assets/img/poster.png', type: PDO::PARAM_STR);
+                $stmtPoster->bindValue(param: ':tipo', value: 'poster', type: PDO::PARAM_STR);
+                $stmtPoster->bindValue(param: ':ordem_exib', value: 0, type: PDO::PARAM_INT);
+                $sucesso = $stmtPoster->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao inserir imagem do poster.");
+                }
 
                 // Define Banner Padrão
                 $sqlBanner = "
@@ -90,10 +121,15 @@
 
                 $stmtBanner = $this->pdo->prepare(query: $sqlBanner);
                 $stmtBanner->bindValue(param: ':id_jogo', value: $idJogo, type: PDO::PARAM_INT);
-                $stmtBanner->bindValue(param: ':caminho', value: 'img/banner.png', type: PDO::PARAM_STR);
+                $stmtBanner->bindValue(param: ':caminho', value: '/assets/img/banner.png', type: PDO::PARAM_STR);
                 $stmtBanner->bindValue(param: ':tipo', value: 'banner', type: PDO::PARAM_STR);
                 $stmtBanner->bindValue(param: ':ordem_exib', value: 0, type: PDO::PARAM_INT);
-                $stmtBanner->execute();
+                $sucesso = $stmtBanner->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao inserir imagem do banner.");
+                }
 
                 $this->pdo->commit();
                 
@@ -128,30 +164,51 @@
                 $sqlJG = "SELECT id_genero FROM jogo_genero WHERE id_jogo = :id";
                 $stmtJG = $this->pdo->prepare(query: $sqlJG);
                 $stmtJG->bindValue(param: ':id',   value: $idJogo, type: PDO::PARAM_INT);
-                $stmtJG->execute();
+                $sucesso = $stmtJG->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao ler gênero do jogo.");
+                }
 
                 $genero = $stmtJG->fetch(PDO::FETCH_ASSOC);
+
                 if (!$genero) {
                     $this->pdo->rollBack(); 
-                    return false; 
+                    return False; 
                 }
 
                 $jogo['generos'] = $genero;
 
-                $sqlLogo = "
-                SELECT caminho 
+                $sqlPoster = "
+                SELECT caminho, id_imagem
                 FROM jogo_imagem 
-                WHERE id_jogo = :id AND tipo = 'logo'
+                WHERE id_jogo = :id AND tipo = 'poster'
                 ";
-                $stmtLogo = $this->pdo->prepare(query: $sqlLogo);
-                $stmtLogo->bindValue(param: ':id', value: $idJogo, type: PDO::PARAM_INT);
-                $stmtLogo->execute();
-                $jogo['logo'] = $stmtLogo->fetchColumn();
+                $stmtPoster = $this->pdo->prepare(query: $sqlPoster);
+                $stmtPoster->bindValue(param: ':id', value: $idJogo, type: PDO::PARAM_INT);
+                $sucesso = $stmtPoster->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao ler imagem do poster.");
+                }
+
+                $jogo['poster'] = $stmtPoster->fetchColumn();
+
+                if (!$jogo['poster']) {
+                    $this->pdo->rollBack();
+                    return False;
+                }
 
                 $sqlBanner = "
-                SELECT caminho 
-                FROM jogo_imagem 
-                WHERE id_jogo = :id AND tipo = 'banner'
+                SELECT 
+                    caminho, 
+                    id_imagem
+                FROM 
+                    jogo_imagem 
+                WHERE 
+                    id_jogo = :id AND tipo = 'banner'
                 ";
 
                 $stmtBanner = $this->pdo->prepare(query: $sqlBanner);
@@ -159,14 +216,20 @@
                 $stmtBanner->execute();
                 $jogo['banner'] = $stmtBanner->fetchColumn();
 
+                if (!$jogo['banner']) {
+                    $this->pdo->rollBack();
+                    return False;
+                }
+
                 $stmtScreenshot = $this->pdo->prepare(query: "
-                SELECT caminho 
+                SELECT 
+                caminho, id_imagem
                 FROM jogo_imagem 
                 WHERE id_jogo = :id AND tipo = 'screenshot'
                 ");
                 $stmtScreenshot->bindValue(param: ':id', value: $idJogo, type: PDO::PARAM_INT);
                 $stmtScreenshot->execute();
-                $jogo['screenshots'] = $stmtScreenshot->fetchAll(PDO::FETCH_COLUMN);
+                $jogo['screenshots'] = $stmtScreenshot->fetchAll(mode: PDO::FETCH_ASSOC);
 
                 $this->pdo->commit();
             
@@ -205,7 +268,12 @@
                 $stmtJogo -> bindValue(param: ":link",              value: $jogo->GetLink(),                type: PDO::PARAM_STR);
                 $stmtJogo -> bindValue(param: ":plataforma",        value: $jogo->GetPlataforma(),          type: PDO::PARAM_STR);
                 $stmtJogo -> bindValue(param: ":id",                value: $jogo->GetId(),                  type: PDO::PARAM_INT);
-                $stmtJogo->execute();
+                $sucesso =$stmtJogo->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao editar jogo.");
+                }
 
                 $sqlJG = "
                 UPDATE jogo_genero
@@ -218,7 +286,12 @@
                 $stmtJG = $this->pdo->prepare(query: $sqlJG);
                 $stmtJG -> bindValue(param: ':id_genero',   value: $jogo->GetGenero(),  type: PDO::PARAM_INT);
                 $stmtJG -> bindValue(param: ':id_jogo',     value: $jogo->GetId(),      type: PDO::PARAM_INT);
-                $stmtJG ->execute();
+                $sucesso =$stmtJG ->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao editar gênero.");
+                }
 
                 $this -> pdo -> commit();
 
@@ -236,6 +309,21 @@
             try {
                 $this->pdo->beginTransaction();
 
+                // Deleta Imagens
+                $sqlImagens = "
+                DELETE FROM jogo_imagem
+                WHERE id_jogo = :id
+                ";
+                $stmtImagens = $this->pdo->prepare(query: $sqlImagens);
+                $stmtImagens->bindValue(param: ":id", value: $id, type: PDO::PARAM_INT);
+                $sucesso = $stmtImagens->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao deletar imagens do jogo.");
+                }
+
+                // Deleta Gêneros
                 $sqlJG = "
                 DELETE FROM jogo_genero
 
@@ -245,8 +333,14 @@
 
                 $stmtJG = $this->pdo->prepare(query: $sqlJG);
                 $stmtJG -> bindValue(param: ":id", value: $id, type: PDO::PARAM_INT);
-                $stmtJG -> execute();
+                $sucesso =$stmtJG -> execute();
 
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao deletar gênero do jogo.");
+                }
+
+                // Deleta Jogo
                 $sqlJogo = "
                 DELETE FROM jogo 
                 
@@ -256,7 +350,12 @@
 
                 $stmtJogo = $this -> pdo -> prepare(query: $sqlJogo);
                 $stmtJogo -> bindValue(param: ":id", value: $id, type: PDO::PARAM_INT);
-                $stmtJogo->execute();
+                $sucesso = $stmtJogo->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao deletar jogo.");
+                }
 
                 $this -> pdo -> commit();
 
@@ -302,16 +401,65 @@
 
                 $jogo['generos'] = $genero;
 
+                $sqlPoster = "
+                SELECT caminho 
+                FROM jogo_imagem 
+                WHERE id_jogo = :id AND tipo = 'poster'
+                ";
+                $stmtPoster = $this->pdo->prepare(query: $sqlPoster);
+                $stmtPoster->bindValue(param: ':id', value: $jogo['id_jogo'], type: PDO::PARAM_INT);
+                $sucesso = $stmtPoster->execute();
+
+                if (!$sucesso) {
+                    $this->pdo->rollBack();
+                    throw new PDOException(message: "Falha ao ler imagem do poster.");
+                }
+
+                $jogo['poster'] = $stmtPoster->fetchColumn();
+
+                if (!$jogo['poster']) {
+                    $this->pdo->rollBack();
+                    return False;
+                }
+
+                $sqlBanner = "
+                SELECT caminho 
+                FROM jogo_imagem 
+                WHERE id_jogo = :id AND tipo = 'banner'
+                ";
+
+                $stmtBanner = $this->pdo->prepare(query: $sqlBanner);
+                $stmtBanner->bindValue(param: ':id', value: $jogo['id_jogo'], type: PDO::PARAM_INT);
+                $stmtBanner->execute();
+                $jogo['banner'] = $stmtBanner->fetchColumn();
+
+                if (!$jogo['banner']) {
+                    $this->pdo->rollBack();
+                    return False;
+                }
+
+                $stmtScreenshot = $this->pdo->prepare(query: "
+                SELECT caminho 
+                FROM jogo_imagem 
+                WHERE id_jogo = :id AND tipo = 'screenshot'
+                ");
+                $stmtScreenshot->bindValue(param: ':id', value: $jogo['id_jogo'], type: PDO::PARAM_INT);
+                $stmtScreenshot->execute();
+                $jogo['screenshots'] = $stmtScreenshot->fetchAll(PDO::FETCH_COLUMN);
+
+                if (!$jogo['screenshots']) {
+                    $this->pdo->rollBack();
+                    return False;
+                }
+
                 $this->pdo->commit();
                 
-
             } catch (\Throwable $e) {
                 if ($this->pdo->inTransaction()) $this->pdo->rollBack();
                 throw new PDOException(message: "Falha ao ler jogo: ".$e->getMessage(), code: 0, previous: $e);
             }
 
             return $jogo;
-
         }
 
         public function CreateImage($idJogo, $caminho, $tipo, $ordem_exib): bool
@@ -359,9 +507,8 @@
 
                 $sqlUpdate = "
                     UPDATE jogo_imagem SET
-                        caminho = :caminho,
-                        tipo = :tipo
-                    WHERE id_jogo = :id_jogo
+                        caminho = :caminho
+                    WHERE id_jogo = :id_jogo AND tipo = :tipo
                 ";
 
                 $stmtUpdate = $this->pdo->prepare(query: $sqlUpdate);
@@ -374,6 +521,26 @@
             } catch (\Throwable $e) {
                 if ($this->pdo->inTransaction()) $this->pdo->rollBack();
                 throw new PDOException(message: "Falha ao atualizar imagem: ".$e->getMessage(), code: 0, previous: $e);
+            }
+
+            return true;
+        }
+
+        public function DeleteImage($id): bool {
+            try {
+                $this->pdo->beginTransaction();
+
+                $sqlDelete = "
+                    DELETE FROM jogo_imagem WHERE id_imagem = :id AND tipo = 'screenshot'
+                ";
+                $stmtDelete = $this->pdo->prepare(query: $sqlDelete);
+                $stmtDelete->bindValue(param: ':id', value: $id, type: PDO::PARAM_INT);
+                $stmtDelete->execute();
+
+                $this->pdo->commit();
+            } catch (\Throwable $e) {
+                if ($this->pdo->inTransaction()) $this->pdo->rollBack();
+                throw new PDOException(message: "Falha ao deletar imagem: ".$e->getMessage(), code: 0, previous: $e);
             }
 
             return true;
